pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-framework-mlx"
version = "0.2.1"
description = "MLX integration for the Agent Framework"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Filip W", email = "contact@strathweb.com" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS :: MacOS X",
]
dependencies = [
    "agent-framework-core",
    "mlx-lm>=0.28.3",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[project.urls]
Homepage = "https://github.com/filipw/agent-framework-mlx"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_framework_mlx"]